refactor(entity): Move entities to Transfers folder
